requirements.txt
django-simple-history
django-currentuser
djangorestframework
rest-framework-generic-relations
djangorestframework_simplejwt